# feat: added method fetchNumberPages that calc number of pages more faster and correctly