## Problem with large file. #138

Imagick too long calculating number of pages and It turned out not always correctly(see test A17_FlightPlan.pdf). Server was down with 503 error on shared hosting because on that max execuution time 30s. I suggest parsing document with Smalot\PdfParser\Parser that work quickly and correctly.

### pdf document 40 pages
old
Time: 37.35 seconds, Memory: 4.00 MB
OK (1 test, 1 assertion)
Process finished with exit code 0
new
Time: 695 ms, Memory: 33.55 MB
OK (1 test, 1 assertion)
Process finished with exit code 0

### pdf document 618 pages
old
Failed asserting that false is true.
#wrong calc number of pages 370 instead of 618
Time: 34.77 seconds, Memory: 6.00 MB
FAILURES!
Tests: 1, Assertions: 1, Failures: 1.

new
Time: 20.79 seconds, Memory: 79.75 MB
OK (1 test, 1 assertion)
Process finished with exit code 0