@rdevroede 2020-01-22
Using pingImage should improve speed, as that only fetches basic info. https://www.php.net/manual/en/imagick.pingimage.php

Something like:
```
$im = new Imagick();
$im->pingImage($pdfFile);
echo $im->getNumberImages();
```

There is no simple pure PHP solution AFAIK.

Other options are to include a dependency like [TCPDI](https://github.com/pauln/tcpdi) or use an external command like pdfinfo.