@rubenvanassche 2020-01-23
Okay I took a look at this, I've tested the whole test suite and these we're the times it took to run the fetch pages function for each test on average:

- PDF Parser: 1.33 seconds
- Current Imagick implementation: 2.70 sconds
- Imagick implementation suggested by @rdevroede:  1.94 seconds

First: @truechernyshov, can you describe me how you could run a test that takes 34.77 seconds to run? When I'm running the current implementation it takes 30.93 seconds for the whole test suite.

Second: The failing test with the bigger pdf file does not crash on my machine, a macOS computer. Haven't tested it on a server deployed by Forge yet though.

Third: The memory usages for the whole test suite are:
- PDF Parser: 79.75 MB
- Current Imagick implementation:  4.00 MB
- Imagick implementation suggested by @rdevroede:  4.00 MB
That's a lot more of memory use for achieving the same task!

Fourth: external tools like `pdfinfo` would require users to install such tools, we want this package to work out of the box and not have to ask users to install extra software when not necessary.

Fifth: I'm not that big of a fan to add new composer dependencies when it can be omitted.

So in conclusion, I think we should replace our Imagick implementation with the one suggested by @rdevroede. The memory usage stays the same, there is a speed bump of 30% and we don't need to add another package to the composer dependencies.

@freekmurze what do you think?
@truechernyshov please could you look a bit more in why the test with number of pages is failing?