@truechernyshov 2020-01-23
Tested on Ubuntu 18.04, php 7.2, phpstorm console
First document was 40pages.pdf, second A17_FlightPlan.pdf
Files attached.
[40pages.pdf](https://github.com/spatie/pdf-to-image/files/4103034/40pages.pdf)
[A17_FlightPlan.pdf](https://github.com/spatie/pdf-to-image/files/4103038/A17_FlightPlan.pdf)