@rubenvanassche 2020-04-29
Wow, this issue completely was lost in my list of issues to do, so I've recorded the times with the two files you provided: 

- PDF Parser: 1.2453881104787
- Current Imagick implementation: 4.6706940134366
- Imagick implementation suggested by @rdevroede: 3.7051098545392

The amount is correct in the tests with the three methods, but memory usage is a lot bigger with `PDF Parser`. I don't see the benefit to add another dependency into the package that isn't our own. When converting such big pdf's a queued approach seems a better solution to this.

I will add  @rdevroede suggestion tho since it is a small fix that will speed up the process a bit.